requirements.txt
# Visualização e análise de dados
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.1
numpy==1.24.3
scipy==1.10.1

# Processamento de imagens
opencv-python==4.7.0.72
Pillow==9.5.0

# Machine Learning e Deep Learning
torch==2.0.0+cpu
torchvision==0.15.1+cpu
kornia==0.6.12
scikit-learn==1.2.2

# Utilitários
tqdm==4.65.0
optuna==3.1.1
ipython==8.12.0